修改 根据最新的配置规则，改写代码生成过程，并修复个别Kotlin重构时的bug

+++ src/main/kotlin/io/github/rcarlosdasilva/cogen/util/namespace.kt
@@ -43,12 +43,15 @@ object Files {
 
   private val logger = KotlinLogging.logger {}
 
-  fun write(content: String, path: String): Boolean {
+  fun write(content: String, path: String, coverage: Boolean): Boolean {
     val file = File(path)
     if (file.isDirectory) {
       return false
     }
 
+    if (file.exists() && !coverage) {
+      return false
+    }
     try {
       FileWriter(file).use { writer ->
         writer.write(content)
@@ -62,12 +65,12 @@ object Files {
   }
 
   fun path(vararg parts: String): String =
-    parts.filter { it.isNotBlank() }.map {
+    parts.filter { it.isNotBlank() }.joinToString("/") {
       val t = it.replace("..", "#")
         .replace('.', '/')
         .replace("#", "..")
       TextHelper.trim(t, "/")
-    }.joinToString { "/" }
+    }
 
 }
